Added GUI to Client and Server
