+++ common/include/common/ui/menu/Menu.h
@@ -38,25 +38,25 @@ public:
     void showWindow(Window window);
     bool isOpen(Window window) const;
     MenuWindow& getWindow(Window window);
-
-private:
+    void makeMenuElements(const std::vector<std::pair<std::string, std::function<void()>>>& elements);
     static void null()
     {
     }
 
+private:
     void draw(sf::RenderTarget& target, sf::RenderStates states) const override;
 
     std::vector<std::pair<std::string, std::function<void()>>> elements_;
-    std::vector<tgui::Button::Ptr> buttons_;
+    Framework* framework_;
+    UserInterface* ui_;
 
+    std::vector<tgui::Button::Ptr> buttons_;
     std::unordered_map<Menu::Window, std::unique_ptr<MenuWindow>> windows_;
 
     AbstractDrawableObject bar_;
     DynamicObject logo_;
     std::list<Event> animation_events_;
 
-    Framework* framework_;
-    UserInterface* ui_;
     tgui::Gui* gui_;
     tgui::Theme* theme_;
