+++ multi/include/server/Server.h
@@ -41,6 +41,7 @@ public:
     void respawn(const std::string& map_name) override;
     void useSpecialObject() override;
     void setGameState(GameState state) override;
+    void startGame(const std::string& map_name) override;
 
 private:
     void beforeUpdate(float time_elapsed) override;
