+++ multi/src/server/MinimalUserInterface.cpp
@@ -2,14 +2,6 @@
 // Created by jul3x on 16.03.21.
 //
 
-#include <iomanip>
-
-#include <R3E/system/Engine.h>
-#include <R3E/utils/Geometry.h>
-#include <R3E/utils/Misc.h>
-
-#include <common/ResourceManager.h>
-
 #include <server/Server.h>
 #include <server/MinimalUserInterface.h>
 
@@ -18,8 +10,38 @@ MinimalUserInterface::MinimalUserInterface(Server* server) : UserInterface(serve
 {
 }
 
+void MinimalUserInterface::initialize(graphics::Graphics& graphics)
+{
+    UserInterface::initialize(graphics);
+    menu_->makeMenuElements({
+            {"Start server", [this]() { menu_->showWindow(Menu::Window::LoadGame); }},
+            {"About",        [this]() { menu_->showWindow(Menu::Window::About); }},
+            {"Exit",         [this]() { framework_->close(); }}
+    });
+
+    full_hud_ = std::make_unique<FullHud>(this, framework_,
+                                          sf::Vector2f{static_cast<float>(CONF<int>("graphics/window_width_px")),
+                                                       static_cast<float>(CONF<int>("graphics/window_height_px"))});
+    menu_->doShow(true);
+}
+
 void MinimalUserInterface::draw(graphics::Graphics& graphics)
 {
+    graphics.setStaticView();
+
+    if (framework_->getGameState() != Framework::GameState::Menu)
+    {
+        graphics.draw(*full_hud_);
+    }
+    else
+    {
+        graphics.draw(*menu_);
+    }
+
+    gui_->draw();
+    graphics.draw(crosshair_);
+
+    RM.setFontsSmoothAllowed(false);
 }
 
 inline void MinimalUserInterface::handleScrolling(float delta)
@@ -32,10 +54,14 @@ inline void MinimalUserInterface::handleKeys()
 
 inline void MinimalUserInterface::handleMouse(sf::RenderWindow& graphics_window)
 {
+    auto mouse_pos = sf::Mouse::getPosition(graphics_window);
+    crosshair_.setPosition(mouse_pos.x, mouse_pos.y);
 }
 
 inline void MinimalUserInterface::updatePlayerStates(float time_elapsed)
 {
+    full_hud_->update(time_elapsed);
+    menu_->update(time_elapsed);
 }
 
 void MinimalUserInterface::spawnAchievement(const std::string& title, const std::string& text, const std::string& tex)
